moved BQT JS into webpack
