+++ app/javascript/packs/legacy.js
@@ -3,7 +3,9 @@ import "isomorphic-fetch";
 import { InitGroupSelection } from "../legacy/group_selection.js";
 import * as legacyBenefitApplications from "../legacy/benefit_application.js";
 import { MetalLevelSelect } from "../legacy/metal_level_select";
+import { Bqt } from "../legacy/bqt";
 InitGroupSelection();
 window.calculateEmployerContributions = legacyBenefitApplications.calculateEmployerContributions;
 window.calculateEmployeeCosts = legacyBenefitApplications.calculateEmployeeCosts;
 global.MetalLevelSelect = MetalLevelSelect;
+global.Bqt = Bqt;
