+++ components/sponsored_benefits/app/views/sponsored_benefits/organizations/plan_design_proposals/plan_selections/new.js.erb
@@ -1,4 +1,4 @@
 setTabContent("<%= escape_javascript render 'new' %>");
 set_active_ontab("#employers-tab");
-pageInit();
+Bqt.loadFunctions();
 window.history.pushState("", "Select a Plan", "<%=  sponsored_benefits.new_organizations_plan_design_proposal_plan_selection_path(@plan_design_proposal.id) %>");
