+++ components/sponsored_benefits/app/views/sponsored_benefits/organizations/plan_design_proposals/plans/_plans.html.erb
@@ -9,7 +9,7 @@
     data-family-deductible="<%= plan.family_deductible %>"
   >
     <div class="col-xs-2">
-      <%= radio_button_tag :reference_plan, "#{plan.id}", false, onclick:"setMyPlans(this)", data:{planId:plan.id} %>
+      <%= radio_button_tag :reference_plan, "#{plan.id}", false, onclick:"Bqt.setMyPlans(this)", data:{planId:plan.id} %>
     </div>
 
     <div class="col-xs-10">
